+++ indiv.h
@@ -33,7 +33,7 @@
 extern void
 rate_super_players 	( bool_t quiet
 					, struct ENC *enc
-					, int N_enc
+					, long N_enc
 					, int *performance_type
 					, int n_players
 					, double *ratingof
